Add new item to the list.
It may seem silly at first, but when you click on that error page after a prerender error and you see that none of the options apply to you, it gets kind of confusing until you actually access the page and see that it's throwing an error, sometimes it's a page that you haven't accessed in a long time and that's why you didn't realize at first. (that's what happened with me). So I think putting that as the first item in the list may save someone else's time in the future.